# Conflicts:
#	car-common/car-common-mapper/src/main/java/com/zb/mapper/AreaMapper.java
#	car-common/car_cars.sql

+++ car-common/car-common-mapper/src/main/java/com/zb/mapper/GarageMapper.java
@@ -8,7 +8,7 @@ import org.apache.ibatis.annotations.Mapper;
 @Mapper
 public interface GarageMapper {
     
-	public Garage getGarageById(@Param(value = "id") Long id)throws Exception;//通过id查询车行
+	public Garage getGarageById(@Param(value = "id") Integer id)throws Exception;//通过id查询车行
 
 	public List<Garage>	getGarageListByMap(Map<String,Object> param)throws Exception;//多条件查询车行
 
@@ -18,7 +18,7 @@ public interface GarageMapper {
 
 	public Integer updateGarage(Garage garage)throws Exception;//更新车行数据
 
-	public Integer deleteGarageById(@Param(value = "id") Long id)throws Exception;//通过id删除车行
+	public Integer deleteGarageById(@Param(value = "id") Integer id)throws Exception;//通过id删除车行
 
 	public Integer batchDeleteGarage(Map<String,List<String>> params);
